requirements.txt
# Movie Scene Replacer Dependencies

# Core AI/ML
openai>=1.0.0
fal-client>=0.5.0  # Veo 3.1 video generation

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0

# Video Processing
scenedetect[opencv]>=0.6.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Script-Subtitle Merging
rapidfuzz>=3.0.0  # Fast fuzzy string matching

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.28.0  # For downloading generated videos